修正GetTickCount越界问题

+++ src/ctp/trader_ctp.h
@@ -73,7 +73,7 @@ private:
     //≤È—Ø«Î«Û
     int ReqQryAccount();
     int ReqQryPosition();
-    int m_next_qry_dt;
+    long long m_next_qry_dt;
     std::atomic_bool m_need_query_positions;
     std::atomic_bool m_need_query_account;
 };
